Put RunCommand into separate file. Set .proto file as origin.
