+++ src/cmd_client.cc
@@ -3,18 +3,6 @@
 
 #include "cmd_client.h"
 
-// Define cmd client flags
-DEFINE_string(registeruser, "", "Registers the given username.");
-DEFINE_string(user, "", "Logs in as the given username.");
-DEFINE_string(caw, "", "Creates a new caw with the given text.");
-DEFINE_string(reply, "", "Indicates that the new caw is a reply to the given id.");
-DEFINE_string(follow, "", "Starts following the given username.");
-DEFINE_string(read, "", "Reads the caw thread starting at the given id.");
-DEFINE_bool(profile, false, "Gets the user’s profile of following and followers.");
-DEFINE_int32(hook, -1, "Specify event type id that will be hooked with event function.");
-DEFINE_int32(unhook, -1, "Specify event type id that will be unhooked.");
-DEFINE_string(event_function, "", "Specify event function.");
-
 void CMDClient::Hook(const int event_type, const std::string &funcstr) {
   if (event_type == -1) {
     std::cout << "event tpye is not specified." << std::endl;
@@ -67,13 +55,12 @@ void CMDClient::RegisterUser(const std::string &username) {
   // Execute registeruser.
   faz::EventReply event_rep = fazclient_.Event(event_req);
 
-  caw::RegisteruserReply rep;
+  caw::RegisteruserRequest rep;
   event_rep.payload().UnpackTo(&rep);
-  if (rep.success()) {
+  if (!rep.username().empty()) {
     std::cout << "Register user " << username << " success." << std::endl;
   }
   else {
-    std::cout << rep.success() << std::endl;
     std::cout << "Register user " << username << " failed." << std::endl;
   }
 }
@@ -145,9 +132,9 @@ void CMDClient::Follow(const std::string &username, const std::string &to_follow
   // Execute registeruser.
   faz::EventReply event_rep = fazclient_.Event(event_req);
 
-  caw::FollowReply rep;
+  caw::RegisteruserRequest rep;
   event_rep.payload().UnpackTo(&rep);
-  if (rep.success()) {
+  if (!rep.username().empty()) {
     std::cout << "Follow user " << to_follow << " success." << std::endl;
   }
   else {
@@ -228,81 +215,4 @@ void PrintCaw(const caw::Caw &c) {
 
 int CMDClient::IsRegistered(const std::string &funcstr) {
   return fazclient_.IsRegistered(funcstr);
-}
-
-void RunCommand() {
-  CMDClient client(grpc::CreateChannel("0.0.0.0:50000",
-                   grpc::InsecureChannelCredentials()));
-  int cmd_count = 0;
-  // Each command should only execute one command.
-  if (!FLAGS_registeruser.empty()) cmd_count++;
-  if (!FLAGS_caw.empty()) cmd_count++;
-  if (!FLAGS_follow.empty()) cmd_count++;
-  if (!FLAGS_read.empty()) cmd_count++;
-  if (FLAGS_profile) cmd_count++;
-  if (FLAGS_hook != -1) cmd_count++;
-  if (FLAGS_unhook != -1) cmd_count++;
-  
-  if (cmd_count > 1) {
-    std::cout << "Command should only execute one command." << std::endl;
-    return;
-  }
-  if (cmd_count == 0) {
-    std::cout << "No executable command found." << std::endl;
-  }
-  // Execute command.
-  if (!FLAGS_registeruser.empty()) {
-    // Check if function is registered.
-    if (client.IsRegistered("registeruser") == -1) {
-      std::cout << "registeruser function is not registered." << std::endl;
-      return;
-    }
-    client.RegisterUser(FLAGS_registeruser);
-  }
-  else if (!FLAGS_caw.empty()) {
-    // Check if function is registered.
-    if (client.IsRegistered("caw") == -1) {
-      std::cout << "caw function is not registered." << std::endl;
-      return;
-    }
-    client.Caw(FLAGS_user, FLAGS_caw, FLAGS_reply);
-  }
-  else if (!FLAGS_follow.empty()) {
-    // Check if function is registered.
-    if (client.IsRegistered("follow") == -1) {
-      std::cout << "follow function is not registered." << std::endl;
-      return;
-    }
-    client.Follow(FLAGS_user, FLAGS_follow);
-  }
-  else if (!FLAGS_read.empty()) {
-    // Check if function is registered.
-    if (client.IsRegistered("read") == -1) {
-      std::cout << "read function is not registered." << std::endl;
-      return;
-    }
-    client.Read(FLAGS_read);
-  }
-  else if (FLAGS_profile) {
-    // Check if function is registered.
-    if (client.IsRegistered("profile") == -1) {
-      std::cout << "profile function is not regitsered." << std::endl;
-      return;
-    }
-    std::cout << client.IsRegistered("profile") << std::endl;
-    client.Profile(FLAGS_user);
-  }
-  else if (FLAGS_hook != -1) {
-    // Hook event type with function.
-    client.Hook(FLAGS_hook, FLAGS_event_function);
-  }
-  else if (FLAGS_unhook != -1) {
-    client.Unhook(FLAGS_unhook);
-  }
-}
-
-int main(int argc, char** argv) {
-  gflags::ParseCommandLineFlags(&argc, &argv, true);
-  RunCommand();
-  return 0;
 }
